requirements.txt
gunicorn==20.0.4
tabulate==0.8.2
timeago==1.0.13
python-telegram-bot==12.6.1
requests==2.20.0